Rename support.font -> support.text
